Moving web to main bot
